refactor: move cloud flag to pro
