+++ pkg/pro/token.go
@@ -1,4 +1,4 @@
-package cloud
+package pro
 
 import (
 	"context"
