+++ pkg/pro/token_test.go
@@ -1,4 +1,4 @@
-package cloud
+package pro
 
 import (
 	"net/http"
